+++ R/shift_funs.R
@@ -10,7 +10,12 @@
 #'  computed for the treatment \code{A}.
 #'
 #' @return A \code{numeric} vector containing the shifted exposure values.
-shift_additive <- function(A, W = NULL, delta) {
-  shifted_treatment <- A + delta
+shift_additive <- function(A, minA, maxA, W = NULL, delta) {
+  if(delta < 0){
+    shifted_treatment <- ifelse((minA<= A + delta),  A + delta, A)
+  } else {
+    shifted_treatment <- ifelse((maxA>= A + delta),  A + delta, A)
+  }
+ 
   return(shifted_treatment)
 }